Update seasons.py axes
When i try to use this function it says there are 6 columns instead of the 5 you specify. Looking at the chart on the website, the only column not addressed is LOG (Length of Game). I only have 6 months of experience so double check me but I'm pretty sure this is the case.